refactor: 이미지 .svg 변경
